requirements.txt
pandas
gspread
gspread-dataframe
google-auth
openpyxl
requests
email-validator
pdfplumber